requirements.txt
Flask~=3.0.3
psycopg2~=2.9.9
pandas~=2.2.2
python-dotenv~=1.0.1
Flask-CORS==4.0.1